AREA Dream Oldale. Items are faster to get. Implemented Hidden Item Quantity. Optimized BgEvent.bgUnion.hiddenItem.hiddenItemId. Renamed Hidden Item Flags and added a count.
merging with 1.3.0

+++ include/global.fieldmap.h
@@ -110,7 +110,7 @@ struct BgEvent
         const u8 *script;
         struct {
             u16 item;
-            u16 hiddenItemId;
+            u8 hiddenItemId;
         } hiddenItem;
         u32 secretBaseId;
     } bgUnion;
